ai_engine/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.1
numpy==1.26.1
scikit-learn==1.3.2